restructured unit tests to add ISA specific folders.
