+++ src/include/simeng/CoreInstance.hh
@@ -23,7 +23,7 @@
 
 // Program used when no executable is provided; counts down from
 // 1024*1024, with an independent `orr` at the start of each branch.
-uint32_t hex_[] = {
+static uint32_t hex_[] = {
     0x320C03E0,  // orr w0, wzr, #1048576
     0x320003E1,  // orr w0, wzr, #1
     0x71000400,  // subs w0, w0, #1
